Combobox bug fix

+++ RoverNetworkManager/Models/RoveCommCustomPacketModel.cs
@@ -7,7 +7,7 @@ namespace RoverNetworkManager.Models
         internal List<string> Commands = new List<string>();
 		internal List<ushort> CommandIDs = new List<ushort>();
 		internal Dictionary<string, string> Addresses = new Dictionary<string, string>();
-		internal ushort SelectedCommand;
+		internal int SelectedCommand;
 		internal string Data;
 		internal string ID;
 		internal string IP;
